third_party/blink/web_tests/platform/generic/printing/iso-page-in-points-expected.txt


PASS: number of pages is 1
All tests passed
